+++ src/Table/Columns/IconColumn/IconColumnSize.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Milenmk\LaravelSimpleDatatables\Table\Columns\IconColumn;
+
+enum IconColumnSize
+{
+    case ExtraSmall;
+
+    case Small;
+
+    case Medium;
+
+    case Large;
+
+    case ExtraLarge;
+
+    case TwoExtraLarge;
+}