+++ CHANGELOG.md
@@ -1,19 +1,31 @@
+## v1.5.0
+
+#### Published at: 2025-03-20
+
+- [NEW] Added IconColumn to Table
+- [NEW] Added heroicons support
+- [NEW] Command to create Livewire component data table support
+
 ## v1.4.2
+
 #### Published at: 2025-03-20
 
 - [FIX] Filters not showing
 
 ## v1.4.1
+
 #### Published at: 2025-03-20
 
 - [FIX] Error Property type not supported in Livewire for property: [{}]
 
 ## v1.4.0
+
 #### Published at: 2025-03-20
 
 - Add filtering functionality
 
 ## v1.3.1
+
 #### Published at: 2025-03-15
 
 - Implement table grouping.
@@ -21,6 +33,7 @@
 - Removed the dedicated Toggle and Checkbox Livewire component. The logic is now inside the HasTable trait.
 
 ## v1.3.0
+
 #### Published at: 2025-03-14
 
 - Added README installation and usage instructions